Move seccomp code to own module
