+++ northstar/Cargo.toml
@@ -70,6 +70,7 @@ npk = [
     "itertools",
     "rand",
     "sha2",
+    "seccomp",
     "serde_with",
     "serde_yaml",
     "tempfile",
@@ -100,6 +101,7 @@ runtime = [
     "url",
     "uuid",
 ]
+seccomp = ["caps", "lazy_static", "memoffset", "nix"]
 
 [build-dependencies]
 anyhow = "1.0"
