+++ northstar/src/npk/manifest.rs
@@ -12,7 +12,10 @@
 //   See the License for the specific language governing permissions and
 //   limitations under the License.
 
-use crate::common::{name::Name, non_null_string::NonNullString, version::Version};
+use crate::{
+    common::{name::Name, non_null_string::NonNullString, version::Version},
+    seccomp::{Seccomp, SyscallRule},
+};
 use itertools::Itertools;
 use serde::{
     de::{Deserializer, Visitor},
@@ -297,49 +300,6 @@ pub enum Mount {
     /// Mount a tmpfs with size
     #[serde(rename = "tmpfs")]
     Tmpfs(Tmpfs),
-}
-
-/// Pre-defined seccomp profiles
-#[derive(Clone, Eq, PartialEq, Debug, Hash, Serialize, Deserialize)]
-pub enum Profile {
-    /// Default seccomp filter similar to docker's default profile
-    #[serde(rename = "default")]
-    Default,
-}
-
-/// Seccomp configuration
-#[derive(Clone, Eq, PartialEq, Debug, Serialize, Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct Seccomp {
-    /// Pre-defined seccomp profile
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub profile: Option<Profile>,
-    /// Explicit list of allowed syscalls
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub allow: Option<HashMap<NonNullString, SyscallRule>>,
-}
-
-#[derive(Clone, Eq, PartialEq, Debug, Serialize, Deserialize)]
-//#[serde(untagged)]
-pub enum SyscallRule {
-    /// Any syscall argument is allowed
-    #[serde(rename = "any")]
-    Any,
-    /// Explicit list of allowed syscalls arguments
-    #[serde(rename = "args")]
-    Args(SyscallArgRule),
-}
-
-#[derive(Clone, Eq, PartialEq, Debug, Serialize, Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct SyscallArgRule {
-    /// Index of syscall argument
-    pub index: usize,
-    /// Value of syscall argument
-    pub values: Option<Vec<u64>>,
-    /// Bitmask of syscall argument
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub mask: Option<u64>,
 }
 
 /// IO configuration for stdin, stdout, stderr
