--- northstar/src/runtime/island/seccomp_profiles/mod.rs
@@ -1,1 +1,0 @@
-pub mod default;
