+++ northstar/src/seccomp/mod.rs
@@ -12,27 +12,13 @@
 //   See the License for the specific language governing permissions and
 //   limitations under the License.
 
-#![deny(clippy::all)]
+// Write Berkeley Packet Filter (BPF) programs
+mod bpf;
+pub use bpf::{seccomp_filter, AllowList};
 
-pub mod common;
+// Predefined seccomp profiles
+pub mod profiles;
 
-#[cfg(feature = "api")]
-/// Northstar remote API. Control start and stop of applications and
-/// receive updates about container states.
-pub mod api;
-
-#[cfg(feature = "npk")]
-/// NPK format support.
-pub mod npk;
-
-#[cfg(feature = "runtime")]
-/// The Northstar runtime core.
-pub mod runtime;
-
-#[cfg(feature = "seccomp")]
-/// Support for seccomp syscall filtering.
-pub mod seccomp;
-
-/// Northstar internal utilities
-#[cfg(feature = "runtime")]
-mod util;
+// internal types
+mod types;
+pub use types::{Profile, Seccomp, SyscallArgRule, SyscallRule};
